Move QbeastOptions to core
