+++ core/src/main/scala/io/qbeast/spark/metadata/MetadataOperation.scala
@@ -15,11 +15,10 @@
  */
 package io.qbeast.spark.metadata
 
-import io.qbeast.core.model.mapper
+import io.qbeast.core.model.{QbeastOptions, mapper}
 import io.qbeast.core.model.Revision
 import io.qbeast.core.model.StagingUtils
 import io.qbeast.core.model.TableChanges
-import io.qbeast.spark.internal.QbeastOptions
 import io.qbeast.spark.utils.MetadataConfig
 import io.qbeast.spark.utils.MetadataConfig.lastRevisionID
 import io.qbeast.spark.utils.MetadataConfig.revision
