+++ src/test/scala/io/qbeast/sources/QbeastOptionsTest.scala
@@ -17,7 +17,7 @@ package io.qbeast.sources
 
 import io.qbeast.core.model.HookInfo
 import io.qbeast.core.model.PreCommitHook.PRE_COMMIT_HOOKS_PREFIX
-import io.qbeast.spark.internal.QbeastOptions
+import io.qbeast.core.model.QbeastOptions
 import io.qbeast.QbeastIntegrationTestSpec
 import org.apache.spark.qbeast.config
 import org.apache.spark.sql.AnalysisException
