+++ src/test/scala/io/qbeast/spark/delta/QbeastSparkTxnTest.scala
@@ -15,7 +15,7 @@
  */
 package io.qbeast.spark.delta
 
-import io.qbeast.spark.internal.QbeastOptions
+import io.qbeast.core.model.QbeastOptions
 import io.qbeast.QbeastIntegrationTestSpec
 import io.qbeast.TestClasses.Student
 import org.apache.spark.sql.delta.DeltaLog
