+++ src/test/scala/io/qbeast/spark/index/SparkRevisionFactoryTest.scala
@@ -23,7 +23,6 @@ import io.qbeast.core.transform.HashTransformer
 import io.qbeast.core.transform.LinearTransformation
 import io.qbeast.core.transform.LinearTransformer
 import io.qbeast.core.transform.ManualPlaceholderTransformation
-import io.qbeast.spark.internal.QbeastOptions
 import io.qbeast.QbeastIntegrationTestSpec
 import io.qbeast.TestClasses.T3
 import org.apache.spark.sql.functions.to_timestamp
