+++ src/test/scala/io/qbeast/spark/writer/WriteTestSpec.scala
@@ -19,7 +19,6 @@ import io.qbeast.core.model._
 import io.qbeast.spark.index.QbeastColumns
 import io.qbeast.spark.index.QbeastColumns._
 import io.qbeast.spark.index.SparkRevisionFactory
-import io.qbeast.spark.internal.QbeastOptions
 import io.qbeast.TestClasses.IndexData
 import org.apache.hadoop.mapreduce.Job
 import org.apache.spark.sql.execution.datasources.parquet.ParquetFileFormat
